Added full IDE-like features for python editing: - <leader>ti to open ipython in split. - full lsp support from coc and coc-python. - <f5> to run current file, <c-f5> to run main.py, :make as well (though not recomended). - bultin packages and installed packages are found in path.
Removed the lexima plugin - It made the whole editting experience slower
and more awkward. It wasnt worth the small gain of the auto closing.

Added a wiki page for cool little know tricks in vim.

Did that Q inside of terminal buffers does :bdelete! so that processes
will be killed.