Made the config possible to run before installing the plugins.
It used to crash and not startup packer on a fresh install because the
plugins' configs would cause errors since the plugins commands dont
exist before their installation.
I changed it so that the plugins' configs are only loaded after the
plugin itself is loaded (except for defx which has a bug it seems).
Also wrapped the line that loads the colorscheme in an error catcher to
avoid that error before the theme's installation.

Moving the plugins a directory will also help in the future with lazy
loading them.

Also started changing some strings from '..' to ".." .